Linting done... along with on-demand import
